Stopped a glitch but created more problems?
With the previous code, if you ran it, after running the first time, it would ask you for a number again and then it would somehow read "0" as if you had typed zero, then, it would loop around back to the beginning and then ask you for the number again, but the second time around, it would work normally. I think this has something to od with the input buffer and limitations of serial.parseINT(). I read online somewhere that if you ended the serial connection and then started it again, it solved the problem and it sort of does, but this feels sort of dodgey.... I can forsee this causing issues in the future. Will have to spend more time to figure out what the problem is.