V2 of rpi code
This code runs a loop which is constantly looking for any inputs from the arduino. If the arduino sends "request" it will make the rpi ask for values in the console. The rpi will ask for commands and it will also ask for angles if you type the command rotate. For every other command, it will send an angle of zero. hence, a command of "set home" will be sent as: set home, 0, 0\n

The current commands availble are: rotate, reset, set home, and status.