Arm module but better
In this version of arm module, the arduino will send "request" to the rpi every time it is ready to take orders. This will cause the rpi to stop request the user input data into the console and then execute accordingly. The arduino has been designed to take serial input in the form of command, angle1, angle2.